+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changelog (most recent first)
 
+- Spreadsheet sample moved from 7GUIs to Miscellaneous. New Cells example that observes more of the specification outlined by 7GUIs (which is mainly about only redrawing/recalculating changes).
 - Spreadsheet example
 - Added tests, but cannot get them to compile (a TODO)
 - Reorganized docs